Chore(.global): move event package
